More robust systemd restart strategy
The path unit can trigger many restarts due to how rsync works. This
might cause sometimes for the units to reach the burst limit and be left
on a failed status even though nothing is wrong with them.

Instead, use a debounce unit that silently restarts by replacing a
transient unit, which is the one that ultimately restarts the relevant
service controlling the webapp, 30 seconds after the last change in the
webapp directory.